Fix a query
OK, so 2 things. First of all, that's not how you write an `ON`. Second, the extensions table doesn't even have an `id` column. This query has never worked. The only reason nobody noticed is because it is run inside of a `try` so the error just gets swallowed. But anyway...wow.